pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cinemaestro"
version = "0.1.0"
description = "Automated short film production pipeline with AI-powered visual generation, audio synthesis, and character consistency"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Cinemaestro Contributors" }]

dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "typer>=0.12,<1.0",
    "httpx>=0.27,<1.0",
    "moviepy>=2.0,<3.0",
    "pydub>=0.25,<1.0",
    "jinja2>=3.1,<4.0",
    "pyyaml>=6.0,<7.0",
    "insightface>=0.7,<1.0",
    "onnxruntime>=1.17,<2.0",
    "numpy>=1.26,<3.0",
    "Pillow>=10.0,<12.0",
    "tqdm>=4.66,<5.0",
    "rich>=13.0,<14.0",
    "tomli>=2.0,<3.0; python_version < '3.11'",
]

[project.optional-dependencies]
cloud = [
    "elevenlabs>=1.0,<2.0",
    "anthropic>=0.40,<1.0",
    "openai>=1.50,<2.0",
    "replicate>=0.30,<2.0",
    "fal-client>=0.4,<1.0",
]
local = [
    "torch>=2.2",
    "diffusers>=0.30",
    "transformers>=4.40",
    "accelerate>=0.30",
    "TTS>=0.22",
]
web = [
    "fastapi>=0.115,<1.0",
    "uvicorn>=0.30,<1.0",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.5",
    "mypy>=1.10",
]
all = ["cinemaestro[cloud,local,web,dev]"]

[project.scripts]
cinemaestro = "cinemaestro.cli:app"

[project.entry-points."cinemaestro.visual_providers"]
# External providers register here

[tool.hatch.build.targets.wheel]
packages = ["src/cinemaestro"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
